populate/requirements.txt
pandas
python-dotenv

#parquet
pyarrow
fastparquet

#database
pymongo